Split PR label modules to satisfy CodeScene file complexity.
Move routing, content rules, and managed label lists into dedicated files so pr-labels.mjs stays a thin facade.
