+++ scripts/lib/pr-label-managed.mjs
@@ -1,0 +1,34 @@
+/** Labels this automation owns (removed before re-applying). */
+export const MANAGED_LABELS = [
+  'promotion',
+  'feature',
+  'fix',
+  'hotfix',
+  'release',
+  'port',
+  'refactor',
+  'integration:primary',
+  'integration:other',
+  'has conflicts',
+  'draft',
+  'ci:pending',
+  'ci:passing',
+  'ci:failing',
+  'changelog:ok',
+  'changelog:missing',
+  'Documentation',
+  'Enhancement',
+  'Bug',
+  'Refactor',
+  'javascript',
+  'dependencies',
+  'quality',
+  'e2e-validated',
+  'porting',
+  'Component: Map',
+  'Component: Outlooks',
+  'Component: Drawing-Tools',
+  'Component: Export',
+  'Component: UI',
+  'Component: Storage',
+];
